perf(control): 消除 f64 与收敛 libm 调用点；重载下 control 230Hz、sensors +43%
- VehicleState.timestamp_s: f64 -> f32（2 处），每拍的 f32<->f64 转换消失
- fdir/vehicle/mission/mavlink/manual/tecs/pid 里 17 处 libm::sqrtf/cosf/fmaxf/fminf
  直调统一收敛到 crate::math::，不再绕过 shim
- math.rs 记录已知开销：libm 0.2 与 micromath 2.1 内部都走 f64（scalbn::<f64>），
  换 crate 无改善；彻底消除 double 需手写纯 f32 多项式近似
- 控制循环加 CTRL_PHASE 分段标记（固定 VMA 0x2001F000，性能剖析用）

实测（重载 setup，arms+uart）：control 227->230Hz，sensors 51->73Hz(+43%)

+++ core/src/hal/sensor.rs
@@ -185,7 +185,7 @@ impl AirspeedSensor for MockAirspeed {
         let noise = crate::math::sin(self.t * 50.0) * 0.05;
         Some(AirspeedSample {
             speed: Airspeed((self.speed + noise).max(0.0)),
-            timestamp_s: self.t as f64,
+            timestamp_s: self.t,
         })
     }
     fn healthy(&self) -> bool { self.healthy }
